+++ CHANGELOG.md
@@ -55,6 +55,7 @@ n'existait deja plus. #1403
 - Correction d'un bug survenant parfois lors de la recuperation d'un fichier dans le cloud causant une exception
 critique. #1384
 - Correction d'un bug empêchant de télécharger les PES en erreur lorsqu'ils étaient trop volumineux.
+- Correction de helios-purge-transaction #1417
 
 ## 5.1.4 - 2025-09-17